+++ tb_plugin/torch_tb_profiler/profiler/communication.py
@@ -1,38 +1,37 @@
 from .. import utils
 logger = utils.get_logger()
 
-class CommunicationParser:
-    def __init__(self):
-        self.comm_node_list = []
 
-    def parse(self, communication_data, steps, steps_names):
-        # Sort the communication node according the start time, this is for correlating communication node between workers
-        for comm_node in communication_data.values():
-            comm_node.kernel_ranges.sort(key=lambda x: (x[0], -x[1]))
-            self.comm_node_list.append(comm_node)
-        self.comm_node_list.sort(key=lambda x: (x.start_time, -x.end_time))
+def generate_communication_nodes(communication_data, steps, steps_names):
+    comm_node_list = []
 
-        # Find each communication node belong to which step
-        index = 0
-        valid_steps = len(steps)
-        for comm_node in self.comm_node_list:
-            while index < valid_steps:
-                if comm_node.start_time >= steps[index][0] and comm_node.end_time <= steps[index][1]:
-                    comm_node.step_name = steps_names[index]
-                    break
-                elif comm_node.start_time >= steps[index][1]:
-                    index += 1
-                else:
-                    logger.error("Found a communication op not belong to any step.")
-                    break
-            if index >= valid_steps:
-                logger.error("Found communication ops not belong to any step. ")
+    # Sort the communication node according the start time, this is for correlating communication node between workers
+    for comm_node in communication_data.values():
+        comm_node.kernel_ranges.sort(key=lambda x: (x[0], -x[1]))
+        comm_node_list.append(comm_node)
+    comm_node_list.sort(key=lambda x: (x.start_time, -x.end_time))
+
+    # Find each communication node belong to which step
+    index = 0
+    valid_steps = len(steps)
+    for comm_node in comm_node_list:
+        while index < valid_steps:
+            if comm_node.start_time >= steps[index][0] and comm_node.end_time <= steps[index][1]:
+                comm_node.step_name = steps_names[index]
                 break
+            elif comm_node.start_time >= steps[index][1]:
+                index += 1
+            else:
+                logger.error("Found a communication op not belong to any step.")
+                break
+        if index >= valid_steps:
+            logger.error("Found communication ops not belong to any step. ")
+            break
 
-        return self.comm_node_list
+    return comm_node_list
 
 
-def aggregate_communication_nodes(comm_node_list):
+def analyze_communication_nodes(comm_node_list):
     step_comm_stats = {}
     total_comm_stats = {}
