+++ tb_plugin/torch_tb_profiler/profiler/data.py
@@ -10,7 +10,7 @@ from collections import OrderedDict
 
 from .. import io, utils
 from . import trace
-from .communication import aggregate_communication_nodes
+from .communication import analyze_communication_nodes
 from .kernel_parser import KernelParser
 from .module_parser import ModuleParser
 from .overall_parser import OverallParser
@@ -142,7 +142,7 @@ class RunProfileData(object):
             self.kernel_stat = kernel_parser.kernel_stat
 
     def communication_parse(self):
-        self.step_comm_stats, self.total_comm_stats = aggregate_communication_nodes(self.comm_node_list)
+        self.step_comm_stats, self.total_comm_stats = analyze_communication_nodes(self.comm_node_list)
 
     def analyze(self):
         self.recommendations = []
